requirement.txt
streamlit
tensorflow
numpy
pillow
pyttsx3
opencv-python
